+++ modules/ppcp-blocks/resources/js/Components/paypal-label.js
@@ -1,0 +1,14 @@
+export const PaypalLabel = ( { components, config } ) => {
+	const { PaymentMethodIcons } = components;
+
+	return (
+		<>
+			<span
+				dangerouslySetInnerHTML={ {
+					__html: config.title,
+				} }
+			/>
+			<PaymentMethodIcons icons={ config.icon } align="right" />
+		</>
+	);
+};
